Fixed rotation issues and displaying mini cooper

+++ app/3D-object-lib/src/main/java/com/colledk/obj3d/parser/data/FaceData.kt
@@ -1,5 +1,5 @@
 package com.colledk.obj3d.parser.data
 
 internal data class FaceData(
-    val vertexIndeces: List<Short>
+    val vertexIndeces: List<Int>
 )
